agent/pyproject.toml
[project]
name = "youplus_future_self"
version = "0.1.0"
description = "YOU+ Future Self accountability agent using Cartesia Line"
requires-python = ">=3.10"
dependencies = [
    "cartesia-line",
    "google-genai>=1.26.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "aiohttp>=3.12.0",
    "uvicorn==0.35.0",
    "supermemory>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]
